Fix datetime calculation for week & month
현재 코드는 월별 및 주별 통계를 낼 때 사용자가 요청한 날짜를 기준으로
월별 및 주별 통계를 보여줍니다. 이 경우 현재 11월인데 11월 16일을
요청하면 문제가 없지만 11월인데 1월을 요청할 경우 1월에 해당하는 주차와
1월에 해당하는 월의 통계만 보여줍니다.

이를 해결하기 위해서는 루프 계산시 오늘 날짜를 기준으로 DateTime 객체
연산을 수행해야 합니다.

+++ bin/opencloset.pl
@@ -6420,6 +6420,18 @@ get '/stat/visitor/:ymd' => sub {
         return;
     }
 
+    my $today = try {
+        DateTime->now(
+            time_zone => app->config->{timezone},
+        );
+    };
+    unless ($today) {
+        app->log->warn("cannot create datetime object: today");
+        $self->redirect_to( $self->url_for('/stat/visitor') );
+        return;
+    }
+    $today->truncate( to => 'day' );
+
     # -2 ~ +2 days from now
     my %count;
     my $from = $dt->clone->truncate( to => 'day' )->add( days => -2 );
@@ -6438,7 +6450,7 @@ get '/stat/visitor/:ymd' => sub {
     }
 
     # from first to current week of this year
-    for ( my $i = $dt->clone->truncate( to => 'year'); $i <= $dt; $i->add( weeks => 1 ) ) {
+    for ( my $i = $today->clone->truncate( to => 'year'); $i <= $today; $i->add( weeks => 1 ) ) {
         my $f = $i->clone->truncate( to => 'week' );
         my $t = $i->clone->truncate( to => 'week' )->add( weeks => 1, seconds => -1 );
 
@@ -6452,7 +6464,7 @@ get '/stat/visitor/:ymd' => sub {
     }
 
     # from january to current months of this year
-    for ( my $i = $dt->clone->truncate( to => 'year'); $i <= $dt; $i->add( months => 1 ) ) {
+    for ( my $i = $today->clone->truncate( to => 'year'); $i <= $today; $i->add( months => 1 ) ) {
         my $f = $i->clone->truncate( to => 'month' );
         my $t = $i->clone->truncate( to => 'month' )->add( months => 1, seconds => -1 );
